+++ src/components/contracts/rpc/src/eth_pubsub.rs
@@ -228,7 +228,7 @@ impl SubscriptionResult {
             let transaction_hash: Option<H256> = if !receipt.logs.is_empty() {
                 Some(H256::from_slice(
                     Keccak256::digest(&rlp::encode(
-                        &block.transactions[receipt_index as usize],
+                        &block.transactions[receipt_index],
                     ))
                     .as_slice(),
                 ))
